refactored fixtures_teardown
