+++ src/functional_tests/extra_utils/util_functions.py
@@ -8,7 +8,7 @@ from django.contrib.auth import (
 )
 from django.contrib.sessions.backends.db import SessionStore
 from django.contrib.auth.models import Permission, User
-
+from django.test.testcases import call_command
 # Additional Libraries
 import hashlib
 
@@ -165,3 +165,29 @@ def generate_session_cookie(usr):
     }
 
     return cookie
+
+
+
+def fixture_teardown(self):
+    # Allow TRUNCATE ... CASCADE and don't emit the post_migrate signal
+    # when flushing only a subset of the apps
+    for db_name in self._databases_names(include_mirrors=False):
+        # Flush the database
+        inhibit_post_migrate = (
+            self.available_apps is not None
+            or (  # Inhibit the post_migrate signal when using serialized
+                # rollback to avoid trying to recreate the serialized data.
+                self.serialized_rollback
+                and hasattr(connections[db_name], "_test_serialized_contents")
+            )
+        )
+        call_command(
+            "flush",
+            verbosity=0,
+            interactive=False,
+            database=db_name,
+            reset_sequences=False,
+            # In the real TransactionTestCase this is conditionally set to False.
+            allow_cascade=True,
+            inhibit_post_migrate=inhibit_post_migrate,
+        )
