+++ src/functional_tests/tests_documentation_rendering.py
@@ -7,7 +7,7 @@ from selenium import webdriver
 from django.test.testcases import call_command, connections
 from django.contrib.staticfiles.testing import LiveServerTestCase
 
-from .extra_utils.util_functions import varnish_cache
+from .extra_utils.util_functions import varnish_cache, fixture_teardown
 from .utils.download_docs import setup_documentation
 import concurrent.futures
 from copy import deepcopy
@@ -18,34 +18,7 @@ from bs4 import BeautifulSoup as Bsoup
 # Core Models
 from .core.models import Version
 
-
-# Fix for CASCADE TRUNCATE FK error
-def _fixture_teardown(self):
-    # Allow TRUNCATE ... CASCADE and don't emit the post_migrate signal
-    # when flushing only a subset of the apps
-    for db_name in self._databases_names(include_mirrors=False):
-        # Flush the database
-        inhibit_post_migrate = (
-            self.available_apps is not None
-            or (  # Inhibit the post_migrate signal when using serialized
-                # rollback to avoid trying to recreate the serialized data.
-                self.serialized_rollback
-                and hasattr(connections[db_name], "_test_serialized_contents")
-            )
-        )
-        call_command(
-            "flush",
-            verbosity=0,
-            interactive=False,
-            database=db_name,
-            reset_sequences=False,
-            # In the real TransactionTestCase this is conditionally set to False.
-            allow_cascade=True,
-            inhibit_post_migrate=inhibit_post_migrate,
-        )
-
-
-LiveServerTestCase._fixture_teardown = _fixture_teardown
+LiveServerTestCase._fixture_teardown = fixture_teardown
 # ---------------------------
