+++ src/functional_tests/tests_events_page.py
@@ -12,7 +12,7 @@ from .core.models import Organisation, OrganisationType
 from .events.models import Event
 
 # Util functions
-from .extra_utils.util_functions import create_unauth_user, create_permitted_user, generate_session_cookie, varnish_cache
+from .extra_utils.util_functions import create_unauth_user, create_permitted_user, generate_session_cookie, varnish_cache, fixture_teardown
 
 
 markup_content = '''
@@ -81,34 +81,7 @@ github: [`Cveinnt/LetsMarkdown.com`](https://github.com/Cveinnt/LetsMarkdown.com
 
 '''
 
-
-# Fix for CASCADE TRUNCATE FK error
-def _fixture_teardown(self):
-    # Allow TRUNCATE ... CASCADE and don't emit the post_migrate signal
-    # when flushing only a subset of the apps
-    for db_name in self._databases_names(include_mirrors=False):
-        # Flush the database
-        inhibit_post_migrate = (
-            self.available_apps is not None
-            or (  # Inhibit the post_migrate signal when using serialized
-                # rollback to avoid trying to recreate the serialized data.
-                self.serialized_rollback
-                and hasattr(connections[db_name], "_test_serialized_contents")
-            )
-        )
-        call_command(
-            "flush",
-            verbosity=0,
-            interactive=False,
-            database=db_name,
-            reset_sequences=False,
-            # In the real TransactionTestCase this is conditionally set to False.
-            allow_cascade=True,
-            inhibit_post_migrate=inhibit_post_migrate,
-        )
-
-
-LiveServerTestCase._fixture_teardown = _fixture_teardown
+LiveServerTestCase._fixture_teardown = fixture_teardown
 # ---------------------------
