+++ src/functional_tests/tests_homepage_links.py
@@ -8,7 +8,7 @@ from selenium import webdriver
 import requests
 from selenium.webdriver.firefox.service import Service
 from django.db import connection
-from .extra_utils.util_functions import varnish_cache
+from .extra_utils.util_functions import varnish_cache, fixture_teardown
 from bs4 import BeautifulSoup as BSoup
 from .extra_utils.crawler import CustomCrawler
 from .utils.download_docs import setup_documentation
@@ -19,33 +19,7 @@ from requests.packages.urllib3.util.retry import Retry
 from .extra_utils.crawler import CustomCrawler
 # Fix for CASCADE TRUNCATE FK error
 
-
-def _fixture_teardown(self):
-    # Allow TRUNCATE ... CASCADE and don't emit the post_migrate signal
-    # when flushing only a subset of the apps
-    for db_name in self._databases_names(include_mirrors=False):
-        # Flush the database
-        inhibit_post_migrate = (
-            self.available_apps is not None
-            or (  # Inhibit the post_migrate signal when using serialized
-                # rollback to avoid trying to recreate the serialized data.
-                self.serialized_rollback
-                and hasattr(connections[db_name], "_test_serialized_contents")
-            )
-        )
-        call_command(
-            "flush",
-            verbosity=0,
-            interactive=False,
-            database=db_name,
-            reset_sequences=False,
-            # In the real TransactionTestCase this is conditionally set to False.
-            allow_cascade=True,
-            inhibit_post_migrate=inhibit_post_migrate,
-        )
-
-
-LiveServerTestCase._fixture_teardown = _fixture_teardown
+LiveServerTestCase._fixture_teardown = fixture_teardown
 # ---------------------------
 
 class CustomLinkTester(LiveServerTestCase):
